Update Standard for Public Code assessment
Most of the changes reflect the minor textual changes that the
Standard for Public Code has undergone.

In one instance a single requirement was split into two,
and the parts of the answers split to match.

Also, we made a couple of small changes we were confident in,
however there may be additional gaps which have already been
closed; those can be part of additional pull requests.

We marked "Document Codebase Maturity" as met, based on
issue #789.
https://github.com/Amsterdam/signals/issues/789

Co-authored-by: Jan Ainali <jan@publiccode.net>