SIG-4683
The end status "GESPLITST" has been used in Amsterdam for a small period.
Signals that have the status "GESPLITS" are basically parent and child
Signals so they also should be deleted after X time.

In this PR the option to delete Signals in the state "GESPLITST" has
been added as an option to the "delete_signals" management command.