SIG-4678
Added the "is_public_accessible" flag of a categroy to the response of
the public category endpoints.

For a parent category the "is_public_accessible" can be set to True or
False. However when there are child categories that have a
"is_public_accessible" set to True this will also apply to the parent
category.